pyproject.toml
[project]
name = "miktos-agentic-system"
version = "0.1.0"
description = "Miktos Agentic System -- Closed-Loop Orchestration Engine"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    "langgraph>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "typing-extensions>=4.0.0",
    "Pillow>=10.0.0",
    "piexif>=1.1.3",
    "obsws-python>=1.7.0",
    "google-api-python-client>=2.0",
    "google-auth-httplib2>=0.2",
    "google-auth-oauthlib>=1.0",
    "requests>=2.31",
    "msal>=1.24",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "types-PyYAML>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "langgraph.*",
    "langchain_core.*",
    "piexif.*",
    "PIL.*",
    "obsws_python.*",
    "pytest",
    "pytest.*",
    "dotenv",
    "dotenv.*",
    "yaml",
    "yaml.*",
    "googleapiclient.*",
    "google.auth.*",
    "google.oauth2.*",
    "google_auth_oauthlib.*",
    "msal",
    "msal.*",
    "requests",
    "requests.*",
    "psutil",
    "psutil.*",
]
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["engine*", "domains*"]
